# GGRC-5031 Optimize permissions query