--- a/src/ggrc_basic_permissions/__init__.py
+++ b/src/ggrc_basic_permissions/__init__.py
@@ -340,6 +340,7 @@ def load_access_control_list(user, permissions):
       func.max(acr.update),
       func.max(acr.delete)
   ).filter(and_(
+      all_models.AccessControlList.object_type != "Relationship",
       all_models.AccessControlList.person_id == user.id,
       all_models.AccessControlList.ac_role_id == acr.id)
   ).group_by(