# Issue description

The loading of user permissions from ACL table took too long.
Two main causes behind that were:
 - slow query because the correct index was not used.
 - slow python code that translated the query results into the permissions dict

# Steps to test the changes

With the latest ggrc-qa db, log in as globcreator and check the speed of responses for any request that requires a full permissions load.

Also pay attention to the object counts and make sure that the user does not lose any access to objects that they previously had.

# Solution description

I have created an index on access control list table that can be used for group_by and added it as a hint to the query.

Then for the slow python code, we have just removed the redundant Relationship objects that were loaded. (we never check permissions on those and those are in the db just for easier propagation)


## Benchmarks (on local machine):
on ggrc-qa db for user "globcreator" with id 402:

generating permissions dict after query executed:
 - with relationships: 0.76969
 - without relationships: 0.55567

Mysql query:
original: 52019 rows in set (1.40 sec)
with index: 52019 rows in set (0.11 sec)
with index without relationships: 34007 rows in set (0.09 sec)

## Overall improvement 

before: ~ 2.2s
after: ~ 0.6s


# Sanity checklist

- [x] I have clicked through the app to make sure my changes work and not break the app.
- [x] I have applied the correct milestone and labels.
- [x] My changes fix the issue described in the description (and do nothing else). 🤞
- [ ] My changes are covered by tests.
- [x] My changes follow our [performance guidelines](https://github.com/google/ggrc-core/blob/dev/docs/source/contributing/performance.rst).
- [ ] My changes follow our [js](https://github.com/google/ggrc-core/blob/dev/docs/source/contributing/javascript.rst) and/or [python](https://github.com/google/ggrc-core/blob/dev/docs/source/contributing/python.rst) guidelines.
- [x] My commits follow our [commit guidelines](https://github.com/google/ggrc-core/blob/dev/docs/source/contributing/git/how_to_write_a_commit_message.rst).

<!-- If your PR includes a migration include the additional checklist items
# Migration checklist
- [ ] Migration passes all checks from our [PR review guidelines](https://github.com/google/ggrc-core/blob/dev/docs/source/contributing/git/reviewing_pull_requests.rst#reviewing-a-pr-containing-database-migration-scripts)
-->

# PR Review checklist

- [x] The changes fix the issue and don't cause any apparent regressions.
- [x] Labels and milestone are correctly set.
- [x] The solution description matches the changes in the code.
- [x] There is no apparent way to improve the performance & design of the new code.
- [x] The pull request is opened against the correct base branch.
- [x] Upon merging, the Jira ticket's fixversion is correctly set and the ticket is moved to "QA - In Progress".

<!-- If your code is not finished yet can include a TODO check list
# TODO

- [ ] First item on the TODO
-->