@NastassiaKiseleva 2018-05-14
@zidarsk8 please add a ticket ggrc-4951 here, also please change a branch to master - as we will be creating a hit fix here 1.10.1