@zidarsk8 2018-05-14
@NastassiaKiseleva This PR is too dangerous to go to master, I have pulled out only the fix needed for the user.

the new PR is here:  https://github.com/google/ggrc-core/pull/7781

The rest of the changes are in this PR.